app/requirements.txt
# Core
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=1.10.2
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.2
psycopg2-binary>=2.9.3

# Development
pytest>=7.2.0
pytest-cov>=4.0.0
black>=23.1.0
isort>=5.12.0
mypy>=1.0.0

# Testing
httpx>=0.23.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0

# Documentation
mkdocs>=1.4.2
mkdocs-material>=9.0.0

# Optional (for async database)
sqlalchemy[asyncio]>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
